cast ap_snprintf() argument to the type expected by the format string (to fix gcc warning on FreeBSD)
git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/trunk@85506 13f79535-47bb-0310-9956-ffa450edef68

+++ modules/filters/mod_include.c
@@ -1196,7 +1196,7 @@ static int handle_fsize(ap_file_t *in, request_rec *r, const char *error, int si
 #if defined(AP_OFF_T_IS_QUAD)
                     ap_snprintf(tag, sizeof(tag), "%qd", finfo.size);
 #else
-                    ap_snprintf(tag, sizeof(tag), "%ld", finfo.size);
+                    ap_snprintf(tag, sizeof(tag), "%ld", (long)finfo.size);
 #endif
                     l = strlen(tag);    /* grrr */
                     for (x = 0; x < l; x++) {
